# api/static/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "API"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/check/{chainId}/{txHash}/{txNonce}": {
      "get": {
        "operationId": "API_CheckWithdrawal",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCheckWithdrawalResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "chainId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "txHash",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "txNonce",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "API_CheckHealth",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "API"
        ]
      }
    },
    "/submit": {
      "post": {
        "operationId": "API_SubmitWithdrawal",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/depositDepositIdentifier"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    }
  },
  "definitions": {
    "apiCheckWithdrawalResponse": {
      "type": "object",
      "properties": {
        "depositIdentifier": {
          "$ref": "#/definitions/depositDepositIdentifier"
        },
        "transferData": {
          "$ref": "#/definitions/depositTransferData"
        },
        "withdrawalStatus": {
          "$ref": "#/definitions/depositWithdrawalStatus"
        },
        "withdrawalIdentifier": {
          "$ref": "#/definitions/depositWithdrawalIdentifier"
        }
      }
    },
    "depositDepositIdentifier": {
      "type": "object",
      "properties": {
        "txHash": {
          "type": "string"
        },
        "txNonce": {
          "type": "string",
          "format": "int64"
        },
        "chainId": {
          "type": "string"
        }
      }
    },
    "depositTransferData": {
      "type": "object",
      "properties": {
        "sender": {
          "type": "string"
        },
        "receiver": {
          "type": "string"
        },
        "depositAmount": {
          "type": "string"
        },
        "withdrawalAmount": {
          "type": "string"
        },
        "commissionAmount": {
          "type": "string"
        },
        "depositAsset": {
          "type": "string"
        },
        "withdrawalAsset": {
          "type": "string"
        },
        "isWrappedAsset": {
          "type": "boolean"
        },
        "depositBlock": {
          "type": "string",
          "format": "int64"
        },
        "signature": {
          "type": "string",
          "title": "used for EVM transfers"
        },
        "merkleProof": {
          "type": "string"
        },
        "txData": {
          "type": "string"
        }
      }
    },
    "depositWithdrawalIdentifier": {
      "type": "object",
      "properties": {
        "txHash": {
          "type": "string"
        },
        "chainId": {
          "type": "string"
        },
        "withdrawalCoreBlock": {
          "type": "string",
          "format": "int64"
        },
        "withdrawalChainBlock": {
          "type": "string",
          "format": "int64"
        },
        "operator": {
          "type": "string"
        }
      }
    },
    "depositWithdrawalStatus": {
      "type": "string",
      "enum": [
        "WITHDRAWAL_STATUS_UNSPECIFIED",
        "WITHDRAWAL_STATUS_PENDING",
        "WITHDRAWAL_STATUS_PROCESSING",
        "WITHDRAWAL_STATUS_SUBMITTING_TO_CORE",
        "WITHDRAWAL_STATUS_PROCESSED",
        "WITHDRAWAL_STATUS_FAILED",
        "WITHDRAWAL_STATUS_ALREADY_WITHDRAWN"
      ],
      "default": "WITHDRAWAL_STATUS_UNSPECIFIED"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com. As of May 2023, there are no widely used type server\nimplementations and no plans to implement one.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}